+++ agentforge4j-llm-vllm/src/main/java/module-info.java
@@ -13,13 +13,11 @@ import com.agentforge4j.llm.vllm.VllmLlmClientFactory;
 module agentforge4j.llm.vllm {
   requires agentforge4j.llm;
   requires agentforge4j.util;
-  requires com.fasterxml.jackson.annotation;
   requires com.fasterxml.jackson.databind;
   requires java.net.http;
   requires org.apache.commons.lang3;
   requires static lombok;
   uses com.agentforge4j.llm.LlmClientFactory;
-  opens com.agentforge4j.llm.vllm.dto to com.fasterxml.jackson.databind;
   provides LlmClientFactory
       with VllmLlmClientFactory;
   provides LlmClientConfigurationAdapter